+++ packages/ckeditor5-dev-tests/bin/prepare-mrgit-json.js
@@ -25,6 +25,10 @@ tools.updateJSONFile( path.join( TEST_DIR_PATH, 'mrgit.json' ), () => {
 		packageName: originalPackageJson.name,
 		// For PR build we want to get the latest commit from given PR instead of Merge Commit.
 		// See: https://github.com/ckeditor/ckeditor5-dev/issues/484
-		commit: process.env.TRAVIS_PULL_REQUEST_SHA || process.env.TRAVIS_COMMIT
+		commit: process.env.TRAVIS_PULL_REQUEST_SHA || process.env.TRAVIS_COMMIT,
+		// Specify a repository that provides the package specified as `packageName` and which should be cloned.
+		// Forked repositories should be able to execute the test scenario as well.
+		// See: https://github.com/ckeditor/ckeditor5-dev/issues/542.
+		repository: process.env.TRAVIS_PULL_REQUEST_SLUG || process.env.TRAVIS_REPO_SLUG
 	} );
 } );