+++ packages/ckeditor5-dev-tests/lib/utils/manual-tests/compilescripts.js
@@ -37,7 +37,8 @@ module.exports = function compileManualTestScripts( options ) {
 		buildDir: options.buildDir,
 		themePath: options.themePath,
 		language: options.language,
-		additionalLanguages: options.additionalLanguages
+		additionalLanguages: options.additionalLanguages,
+		debug: options.debug
 	} );
 
 	return runWebpack( webpackConfig );